[loader] menu.c: Print a warning if the game list couldn't be loaded.
If /games/ is missing, booting from disc is still allowed on Wii.
If the device is missing, the menu still allows the user to go back.

This should help for those times when you accidentally select the wrong
device, resulting in Nintendont exiting to the loader instead of letting
you go back and select the correct device.