[libcustomfat] partition.c: Ignore EFI system partitions.
Mac OS X seems to create an EFI system partition on all GPT and/or
Hybrid MBR disks, even if it isn't being used to boot the system.

On Hybrid MBR, these partitions aren't present in the MBR, but they
are present in the GPT.

Ignore them when searching for the Nintendont partition.

Reference: https://github.com/FIX94/Nintendont/issues/202