[kernel] ff.c: Added support for MBR extended partitions.
This allows Nintendont to load games from a logical partition. Note that
it's still limited to the first FAT partition found on the device, so
this will work in the relatively common case of an NTFS partition as
primary partition 1, then a FAT32 partition as a logical partition.

NOTE: Only partitions with type 0x0F (LBA extended) are supported.
Linux fdisk and gparted (and possibly some other utilities) create
extended partitions with type 0x05 (CHS extended).