[fatfs] Enable standard disk read/write functions on PowerPC.
Switching over to device numbers in the kernel (ARM) caused some
weird lag issues to show up in Melee, sometimes causing an error
message to show up.

Using a single volume with the kernel works fine, and since we
aren't planning on switching between SD and USB at runtime, this
shouldn't be an issue.

Enable 2 volumes with string device names on PowerPC, since this
will be used by the loader later on.