Use -std=gnu99 in all projects; use -O3 in the loader.
libfatfs-ppc.a was failing to build due to ff_utf8.c using a for loop
variable initialization, e.g. for (int i = blah; ;), which is valid
in C99 but not C89/C90.

grrlib.c: Fixed some compilation errors in gnu99 mode.
TODO: -fgnu89-inline is currently needed for grrlib.h.