[loader] diskio.c: Ported over FIX94's basic diskio cache.
This will only be implemented in the loader, since it has to check
lots of files and parse directories. The kernel only accesses a
couple of files during normal operation, e.g. the disc image and
the emulated memory card.

Sidenote: I've noticed that device initialization seems to be
slightly slower than with libfat/libcustomfat (about a second
or so). It might be possible to optimize this later.