[loader] diskio.c: Fixed a MASSIVE DATA CORRUPTION BUG in the caching code.
This data corruption bug occured if an SD card was used and a USB hard
drive was *not* connected. Basically, it would result in the caching
code using a sector size of 0, since it used usbstorage.c's
__sector_size value, which isn't initialized if no USB hard drive
is connected. The cache would therefore have values set to indicate
it had "valid" data, but the actual data blocks would be garbage.

This resulted in my test SD card getting corrupted multiple times.

Solution: Added a variable to the cache struct indicating the actual
sector size of the device. If it's USB, use __sector_size; otherwise,
hardcode 512. Also added a santiy check to ensure that sectorSize
is >= 512 before caching any sectors.