[loader] Colorize games based on their file format.
GameInfo's NameAlloc field is now Flags, and contains the name allocation
flag as well as the format:
- 0: Full 1:1 rip
- 1: Shrunken
- 2: Extracted FST
- 3: CISO format

"Shrunken" is determined by checking if the game is a normal GCN game.
If it is, and its size is not 1,459,978,240, then it's assumed to have
been shrunken using a tool like DiscEX.

TODO: Triforce games are smaller than GCN discs, so we'll need to add
exceptions for those.