[loader] menu.c: Detect multi-game discs and display them in purple.
Use a custom name instead of titles.txt, since e.g. GCO?DV collides with
"Call of Duty: Finest Hour". The custom name contains the filename as
well, so you can have multiple multi-game disc images on the same
hard drive and distinguish between them if you really want to.

NOTE: While it is technically possible to support multi-game discs in
CISO format, we're not doing that because it requires more work in the
loader, and if you're using CISO, you might as well split the image
into individual files per game.