[loader] main.c: Split multi-game disc selection into a separate function.
New function CheckForMultiGame() checks if this is a multi-game disc.
If not, it returns immediately with an ISOShift value of 0. Otherwise,
it prompts the user to select one of the games, and then returns that
game's offset.

Added a prompt to select a game from the multi-game disc.

TODO: Use menu.c's menu code.

Split multi-game disc checking into a new function in global.c,
IsMultiGameDisc(). Updated main.c and menu.c to use this function.