[loader] titles.c: Optimized TriforceTitles[].
titleID[] should be at least 7 characters in order to have a NULL
terminator. Let's make it 8 so it's nice and round.

Since most of the actual titles are duplicated (one per region), use
const char* instead of const char[]. The duplicate entries will all
point to the same titleName string.