[loader] CheckForMultiGameAndRegion() wasn't returning the correct value for multi-game images.
Since the bi2.bin code was added, the return value should be 0 for
success and non-zero for error, not the ISOShift.

Fixed the function to set ISOShift and return 0 on success.

main(): Initialize ISOShift and BI2region to 0.

This fixes a regression from commit 4d8a6538d98e51571b8a28943373cf4c0fd2956d.
([loader] Use the bi2.bin region code to set the system region.)