[loader] Ported libfat's caching code to FatFS for use in the loader.
This eliminates the memory card initialization lag I saw before,
which was ~6-7s on my setup. It might also fix similar issues on
other systems, where it could hang for 5 minutes or more.

The issue was the previous caching code didn't do any write caching.
Any writes caused the relevant sectors to be removed from the cache.
When creating the file, this results in the FAT being written to
numerous times, so the lack of proper write-back caching caused it
to keep re-reading and re-writing FAT blocks.