GameTDB-md5-parser.pl: Support multi-disc games.
Added $record->{'md5s'}, which uses the game version as the key
and the MD5 as the value. This allows for multi-disc games to
work properly.

Expanded the DB format to include the disc number. The Nintendont
verifier doesn't support this right now, so it will show the disc
number as part of the title.

TODO: Parse the version field for multi-disc in case it actually
has a version number in addition to a disc number.