[loader] update.c: Added gcn_md5.txt.
gcn_md5.txt is compressed as a Zip file in the repository, so it has
to be decompressed. Note that unlike controllers.zip, gcn_md5.txt
must stay on the loader's drive, not the "active" drive.

downloads_t: Changed string arrays to pointers. The arrays aren't nearly
full, so there's no point in wasting space just to reduce pointer usage.