[loader] verify_md5.c: Show the disc number; use the header for ID6 and revision.
GameTDB has the same MD5 for a few discs that shipped in boxes with
different game IDs but have the same contents. Use the disc header
instead of the database ID6 and revision for these.

Disc number is always taken from the database, since the header doesn't
distinguish between "disc 1" and "single-disc game".

This basically wraps up the MD5 verifier.