[loader] ShowGameInfo.c: Initial support for MD5 calculation.
"Calculate MD5" is always shown as the "A" action, but is grayed out
if the disc image isn't 1:1.

If the database couldn't be loaded, an error is shown. The MD5 can
still be calculated (assuming it's 1:1), but it won't be verified.

TODO: Port the actual MD5 verification from verify_md5.c.