add steam controller support
Steam controllers use signed little-endian shorts for the thumbstick and virtual thumbstick. We will use the big end as a signed byte.
NOTE: I do NOT know if the axis direction is correct - I assume a positive Y for up and a positive X for right. Also, I am not including the configuration file for the steam controller.