Adding RE Remake, Zero, 2 & 3 to Widescreen excep.
Included all of the first four Resident Evil titles to the Force Widescreen exception list.
All of these four games are incompatible with the Force Widescreen patch. 
If enabled, it creates an odd side-effect in which the 3D objects are properly skinned down, but the prerendered images; being 4:3 natively, are stretch out.
This causes the 3D objects to be out of place in all of the game's scenes.

If needed/wanted, I could work on also adding games which don't play nice with the Force WIdescreen option.