delay networking initialization until after Melee boots
* Get rid of early network initialization in the loader (via libogc)
* Translate libogc network init into NCDInit(), to-be-run directly on Starlet
* Use the timers in kernel/main.c to serialize some things:
	* Start networking some ~30s after we boot into the new kernel
	* Start our server thread some ~5s after networking

When (a) using libogc to initialize networking in the loader, and when (b)
doing networking operations (like binding to a socket) immediately before
booting: Melee seems to crash very early.

We suspect that networking threads are more-often-than-not on-CPU when Melee
really _needs_ some other work to be done in order to actually boot (i.e. in
order to service disc reads, or something like that).