add support for recovering from overwrite reads
before there was a bug where if a network disconnected for a while, or connected late, the read pointer would be way behind the write cursor of slippi memory. This means that the data it was trying to read was already overwritten. This code allows for the reader to reset to the write cursor if an overflow read is detected