F-Zero AX region change
The "SystemRegion" value was changed from "REGION_JAPAN" to "REGION_EXPORT" and the memory address change to force the language to English was commented. This is a more elegant way to switch the language displayed in F-Zero AX in English. Some text will become more legible, too.
You can also use the REGION_USA value as well, which will add a "Parental Advisory Warning" message, although it is kind of annoying IMO.