handle the dance mat case
https://github.com/FIX94/Nintendont/issues/527

is solved by this commit.

I needed to pull in the changes from here: https://github.com/FIX94/Nintendont/pull/778
to get it to build.

in my environment (for reference) I needed to add a couple things to PATH:
/cygdrive/c/devkitPro/devkitPPC/bin:/cygdrive/c/devkitPro/tools/bin

I tried launching from usbloadergx but for some reason wont loader
but loading nintendont from the homebrew channel and launching Mario Mix
this solves the problem - so I am quite certain this is the solution and
it is possible. I am guessing the crash when loading from usbloadergx
has more to do with the build environment. I had a hard time getting it to build.
The actual code changes to make it work are quite small but the generated dol is about a MB larger, than the existing official one,
so i am guessing devkit has a bunch of stuff that wasnt in the old version.

The fix is here, it just needs a little work to be integrated by someone who knows the code and build environment much better

Added ability for dance pads to exit back to HBC by pressing the four
corner buttons simultaneously. X, Y, A, B. this requires 2 feet and
2 hands and is unlikely to happen by unintentionally.